# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: รินจัดให้ - Shop API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a single service with its packages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service with packages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceWithPackages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateService
      tags: [services]
      summary: Update a service (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── PACKAGES ────────────────────────────────────────────────
...
  /payments/slip:
    post:
      operationId: uploadSlip
      tags: [payments]
      summary: Upload payment slip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SlipInput"
      responses:
        "201":
          description: Slip uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}/verify:
    put:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify payment (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentVerifyInput"
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # ─── GALLERY ─────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    # ─── AUTH ───────────────────────────────────────────────────
...
    ServiceInput:
      type: object
      required: [name, description, icon, category]
      properties:
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
    Service:
      type: object
      required: [id, name, description, icon, category, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ServiceWithPackages:
      type: object
      required: [id, name, description, icon, category, isActive, createdAt, packages]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        packages:
          type: array
          items:
            $ref: "#/components/schemas/Package"

    # ─── PACKAGES ────────────────────────────────────────────────
...
    Package:
      type: object
      required: [id, serviceId, name, price, durationDays, features, isPopular, createdAt]
      properties:
        id:
          type: integer
        serviceId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        durationDays:
          type: integer
        features:
          type: array
          items:
            type: string
        isPopular:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # ─── ORDERS ──────────────────────────────────────────────────
...
    SlipInput:
      type: object
      required: [orderId, slipImageBase64]
      properties:
        orderId:
          type: integer
        slipImageBase64:
          type: string
        amount:
          type: number
    PaymentVerifyInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [verified, rejected]
        note:
          type: string
    Payment:
      type: object
      required: [id, orderId, slipImageBase64, status, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        slipImageBase64:
          type: string
        amount:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, verified, rejected]
        note:
          type: ["string", "null"]
        verifiedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time

    # ─── GALLERY ─────────────────────────────────────────────────
...