# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * รินจัดให้ - Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  name: string;
  description: string;
  icon: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  isActive: boolean;
  createdAt: Date;
}